Use panic() in the type checks
We need to do our own type checks so that the Go runtime will not try
to treat one of our immediate objects as an interface{}. Once we are
sure that a value is heap-allocated we can let the runtime check if
it's a vector, a pair, etc. Later the VM will have enough information
to print error messages nicer than "bad type".