Add a compile-bytecode procedure and bytecode support in load
It's now possible to compile a script to bytecode:

  (bytecode-compile "tests.scm")

The compiled file in the example is tests.cso. When loading Scheme
code the extension is first changed to .cso and an attempt is made to
load the bytecode:

  (load "tests.scm")

This works even if tests.scm is deleted.

Also fixes bugs in memq, memv and member. They did not return the
list.

Fixes issue #4.